Re-implementing slider.
