+++ src/template.html
@@ -3,9 +3,14 @@
   <head>
     <meta http-equiv="Content-type" content="text/html; charset=utf-8"/>
     <title></title>
+    <link rel="stylesheet" href="http://code.jquery.com/ui/1.11.4/themes/redmond/jquery-ui.css">
   </head>
   <body>
     <div id="map" style="width: 100%; height: 100vh;"></div>
+    <div id="day-panel">
+        <div id="day-label"></div>
+        <div id="day-slider"></div>
+    </div>
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.12.2/jquery.min.js"></script>
     <script src="//code.jquery.com/ui/1.11.4/jquery-ui.js"></script>
   </body>
